:beers: finish UART TX experiment

+++ Project/src/Timer.h
@@ -3,8 +3,9 @@
 
 #include "global.h"
 
-void DelayUsec(uchar num);
-void DelayMsec(uchar num);
+// use Timer 1
+void DelayUsec(short int num);
+void DelayMsec(int num);
 
 #endif // !TIMER
